add: offer controller and calculation offer

+++ PowerBlog.Site/Models/OfferPay.cs
@@ -8,8 +8,8 @@ namespace PowerBlog.Site.Models
         public long Id { get; set; }
         [Required]
         public string? OfferWord { get; set; }
-        public double? OfferPercentage { get; set; }
-        public double? OfferAmount { get; set; }
+        public int? OfferPercentage { get; set; }
+        public decimal? OfferAmount { get; set; }
         public string? Description { get; set; }
         public bool IsDisable { get; set; }
         public DateTime? CreatDate { get; set; }
